+++ src/Command.php
@@ -6,8 +6,7 @@ namespace Yiisoft\Db\Mysql;
 
 use PDOException;
 use Throwable;
-use Yiisoft\Db\Driver\PDO\AbstractCommandPDO;
-use Yiisoft\Db\Driver\PDO\ConnectionPDOInterface;
+use Yiisoft\Db\Driver\Pdo\ConnectionInterface;
 use Yiisoft\Db\Exception\ConvertException;
 use Yiisoft\Db\QueryBuilder\QueryBuilderInterface;
 
@@ -15,7 +14,7 @@ use Yiisoft\Db\QueryBuilder\QueryBuilderInterface;
  * Implements a database command that can be executed with a PDO (PHP Data Object) database connection for MySQL,
  * MariaDB.
  */
-final class Command extends AbstractCommandPDO
+final class Command extends \Yiisoft\Db\Driver\Pdo\AbstractCommand
 {
     public function insertWithReturningPks(string $table, array $columns): bool|array
     {
@@ -75,7 +74,7 @@ final class Command extends AbstractCommandPDO
                     && $this->db->getTransaction() === null
                 ) {
                     $this->db->transaction(
-                        fn (ConnectionPDOInterface $db) => $this->internalExecute($rawSql),
+                        fn (ConnectionInterface $db) => $this->internalExecute($rawSql),
                         $this->isolationLevel,
                     );
                 } else {
