+++ src/Schema.php
@@ -9,7 +9,6 @@ use Throwable;
 use Yiisoft\Db\Constraint\Constraint;
 use Yiisoft\Db\Constraint\ForeignKeyConstraint;
 use Yiisoft\Db\Constraint\IndexConstraint;
-use Yiisoft\Db\Driver\PDO\PdoAbstractSchema;
 use Yiisoft\Db\Exception\Exception;
 use Yiisoft\Db\Exception\InvalidConfigException;
 use Yiisoft\Db\Exception\NotSupportedException;
@@ -92,7 +91,7 @@ use function trim;
  *   }
  * >
  */
-final class Schema extends PdoAbstractSchema
+final class Schema extends \Yiisoft\Db\Driver\Pdo\AbstractSchema
 {
     /**
      * @var array Mapping from physical column types (keys) to abstract column types (values).
