+++ src/Transaction.php
@@ -4,11 +4,9 @@ declare(strict_types=1);
 
 namespace Yiisoft\Db\Mysql;
 
-use Yiisoft\Db\Driver\PDO\AbstractTransactionPDO;
-
 /**
  * Implements the MySQL, MariaDB specific transaction.
  */
-final class Transaction extends AbstractTransactionPDO
+final class Transaction extends \Yiisoft\Db\Driver\Pdo\AbstractTransaction
 {
 }
