+++ tests/Pdo/CommandTest.php
@@ -2,17 +2,16 @@
 
 declare(strict_types=1);
 
-namespace Yiisoft\Db\Mysql\Tests;
+namespace Yiisoft\Db\Mysql\Tests\Pdo;
 
 use Yiisoft\Db\Mysql\Tests\Support\TestTrait;
-use Yiisoft\Db\Tests\Common\CommonCommandPDOTest;
 
 /**
  * @group mysql
  *
  * @psalm-suppress PropertyNotSetInConstructor
  */
-final class CommandPDOTest extends CommonCommandPDOTest
+final class CommandTest extends \Yiisoft\Db\Tests\Common\Pdo\CommonCommandTest
 {
     use TestTrait;
 }
