+++ tests/Support/TestTrait.php
@@ -4,7 +4,7 @@ declare(strict_types=1);
 
 namespace Yiisoft\Db\Mysql\Tests\Support;
 
-use Yiisoft\Db\Driver\PDO\ConnectionPDOInterface;
+use Yiisoft\Db\Driver\Pdo\ConnectionInterface;
 use Yiisoft\Db\Mysql\Connection;
 use Yiisoft\Db\Mysql\Dsn;
 use Yiisoft\Db\Mysql\Driver;
@@ -14,7 +14,7 @@ trait TestTrait
 {
     private string $dsn = '';
 
-    protected function getConnection(bool $fixture = false): ConnectionPDOInterface
+    protected function getConnection(bool $fixture = false): ConnectionInterface
     {
         $db = new Connection(new Driver($this->getDsn(), 'root', ''), DbHelper::getSchemaCache());
 
@@ -25,7 +25,7 @@ trait TestTrait
         return $db;
     }
 
-    protected static function getDb(): ConnectionPDOInterface
+    protected static function getDb(): ConnectionInterface
     {
         $dsn = (new Dsn('mysql', '127.0.0.1', 'yiitest', '3306', ['charset' => 'utf8mb4']))->asString();
